refactor: Update search functionality in Store component
The code changes in `Store.php` update the search functionality in the `Store` component. The `$q` property has been renamed to `$search` for clarity. The `mount()` method has been updated to use the new `$search` property instead of `$q` when querying the products. Additionally, the `updatedSearch()` method has been modified to reset the `tag` and `category` properties and update the products based on the new search value.

Note: This commit message follows the established convention of using a verb in the imperative mood to describe the purpose of the code changes.